// src/app/Modules/ViewContact/Events/types.ts
export type TElements = {
  button: HTMLElement | null;
  viewContact: HTMLElement | null;
} | null;

export type TCustomClassesElements = {
  closeViewContactMobile?: "js-close-view-contact-mobile";
  openViewContactMobile?: "js-open-view-contact-mobile";
  containerViewContact?: "js-container-view-contact";
  closeViewContact?: "js-close-view-contact";
  openViewContact?: "js-open-view-contact";
};